new partial for lib-name bits
