+++ src/StackExchange.Redis/ConnectionMultiplexer.LibraryName.cs
@@ -1,0 +1,93 @@
+﻿using System;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading;
+
+namespace StackExchange.Redis;
+
+public partial class ConnectionMultiplexer
+{
+    private string[] libraryNameSuffixes = Array.Empty<string>();
+
+    /// <summary>
+    /// Append a usage-specific modifier to the advertised library name; suffixes are de-duplicated
+    /// and sorted alphabetically. Connections will be updated as necessary (RESP2 subscription
+    /// connections will not show updates until those connections next connect).
+    /// </summary>
+    public void AddLibraryNameSuffix(string suffix)
+    {
+        if (string.IsNullOrWhiteSpace(suffix)) return; // trivial
+
+        // sanitize and re-check
+        suffix = ServerEndPoint.ClientInfoSanitize(suffix ?? "").Trim();
+        if (string.IsNullOrWhiteSpace(suffix)) return; // trivial
+
+        string[] oldValue, newValue;
+        do
+        {
+            oldValue = Volatile.Read(ref libraryNameSuffixes);
+            if (oldValue.Contains(suffix)) return; // already cited
+
+            // otherwise, create a new extended array with the new value (sorted)
+            newValue = new string[oldValue.Length + 1];
+            oldValue.CopyTo(newValue, 0);
+            newValue[newValue.Length - 1] = suffix;
+            Array.Sort(newValue);
+
+            // swap the field (CEX), or redo from start; note this API is very low usage, and is unlikely to ever conflict
+        } while (!ReferenceEquals(Interlocked.CompareExchange(ref libraryNameSuffixes, newValue, oldValue), oldValue));
+
+        // if we get here, we *actually changed something*; we can retroactively fixup the connections
+        var libName = GetFullLibraryName(); // note this also checks SetClientLibrary
+        if (string.IsNullOrWhiteSpace(libName) || !CommandMap.IsAvailable(RedisCommand.CLIENT)) return; // disabled on no lib name
+
+        // note that during initial handshake we use raw Message; this is low frequency - no
+        // concern over overhead of Execute here
+        var args = new object[] { RedisLiterals.SETINFO, RedisLiterals.lib_name, libName };
+        foreach (var server in GetServers())
+        {
+            try
+            {
+                // note we can only fixup the *interactive* channel; that's tolerable here
+                if (server.IsConnected)
+                {
+                    // best effort only
+                    server.Execute("CLIENT", args, CommandFlags.FireAndForget);
+                }
+            }
+            catch (Exception ex)
+            {
+                // if an individual server trips, that's fine - best effort; note we're using
+                // F+F here anyway, so we don't *expect* any failures
+                Debug.WriteLine(ex.Message);
+            }
+        }
+    }
+
+    internal string GetFullLibraryName()
+    {
+        var config = RawConfig;
+        if (!config.SetClientLibrary) return ""; // disabled
+
+        var libName = config.LibraryName;
+        if (string.IsNullOrWhiteSpace(libName))
+        {
+            // defer to provider if missing (note re null vs blank; if caller wants to disable
+            // it, they should set SetClientLibrary to false, not set the name to empty string)
+            libName = config.Defaults.LibraryName;
+        }
+
+        libName = ServerEndPoint.ClientInfoSanitize(libName);
+        // if no primary name, return nothing, even if suffixes exist
+        if (string.IsNullOrWhiteSpace(libName)) return "";
+
+        // append any suffixes (note they're already sanitized)
+        var suffixes = Volatile.Read(ref libraryNameSuffixes);
+        if (suffixes is { Length: > 0 })
+        {
+            libName += "-" + string.Join("-", suffixes);
+        }
+
+        return libName;
+    }
+}
